Remove quick filter option and rename folder User Registration
