+++ frontend/src/pages/UserRegistration/UserRegistrationColumns.tsx
